// lib/li.fi-ts/models/Bridge.ts
/**
 * LI.FI API
 * LI.FI provides the best cross-chain swap across all liquidity pools and bridges.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BridgesEnum } from './BridgesEnum';
import { SupportedChains } from './SupportedChains';
import { HttpFile } from '../http/http';

export class Bridge {
    'key'?: BridgesEnum;
    /**
    * The common name of the tool
    */
    'name'?: string;
    /**
    * The logo of the tool
    */
    'logoURI'?: string;
    'supportedChains'?: Array<SupportedChains>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{ name: string, baseName: string, type: string, format: string }> = [
        {
            "name": "key",
            "baseName": "key",
            "type": "BridgesEnum",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "logoURI",
            "baseName": "logoURI",
            "type": "string",
            "format": ""
        },
        {
            "name": "supportedChains",
            "baseName": "supportedChains",
            "type": "Array<SupportedChains>",
            "format": ""
        }];

    static getAttributeTypeMap() {
        return Bridge.attributeTypeMap;
    }

    public constructor() {
    }
}