// lib/li.fi-ts/models/BridgeDefinition.ts
/**
 * LI.FI API
 * LI.FI provides the best cross-chain swap across all liquidity pools and bridges.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { BridgesEnum } from './BridgesEnum';
import { Token } from './Token';
import { HttpFile } from '../http/http';

/**
* The defintion of a bridge with the transferrable token pair
*/
export class BridgeDefinition {
    'tool': BridgesEnum;
    /**
    * From which chain transfers can happen
    */
    'fromChainId': number;
    'fromToken': Token;
    /**
    * To which chain transfers can happen
    */
    'toChainId': number;
    'toToken': Token;
    /**
    * The maximum amount that can be sent in one transfer
    */
    'maximumTransfer': string;
    /**
    * The minimum amount that can be sent in one transfer
    */
    'minimumTransfer': string;
    /**
    * The rate taken for swap fees
    */
    'swapFeeRate': string;
    /**
    * The minimum swap fees that will be taken
    */
    'swapFeeMinimum': string;
    /**
    * The maximum swap fees that might occur
    */
    'swapFeeMaximum': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{ name: string, baseName: string, type: string, format: string }> = [
        {
            "name": "tool",
            "baseName": "tool",
            "type": "BridgesEnum",
            "format": ""
        },
        {
            "name": "fromChainId",
            "baseName": "fromChainId",
            "type": "number",
            "format": "number"
        },
        {
            "name": "fromToken",
            "baseName": "fromToken",
            "type": "Token",
            "format": ""
        },
        {
            "name": "toChainId",
            "baseName": "toChainId",
            "type": "number",
            "format": "number"
        },
        {
            "name": "toToken",
            "baseName": "toToken",
            "type": "Token",
            "format": ""
        },
        {
            "name": "maximumTransfer",
            "baseName": "maximumTransfer",
            "type": "string",
            "format": ""
        },
        {
            "name": "minimumTransfer",
            "baseName": "minimumTransfer",
            "type": "string",
            "format": ""
        },
        {
            "name": "swapFeeRate",
            "baseName": "swapFeeRate",
            "type": "string",
            "format": ""
        },
        {
            "name": "swapFeeMinimum",
            "baseName": "swapFeeMinimum",
            "type": "string",
            "format": ""
        },
        {
            "name": "swapFeeMaximum",
            "baseName": "swapFeeMaximum",
            "type": "string",
            "format": ""
        }];

    static getAttributeTypeMap() {
        return BridgeDefinition.attributeTypeMap;
    }

    public constructor() {
    }
}