// lib/li.fi-ts/models/Chain.ts
/**
 * LI.FI API
 * LI.FI provides the best cross-chain swap across all liquidity pools and bridges.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ChainMetamask } from './ChainMetamask';
import { Token } from './Token';
import { HttpFile } from '../http/http';

/**
* Representation of a chain
*/
export class Chain {
    /**
    * Short string represenation of the chain
    */
    'key': string;
    /**
    * Type of the chain
    */
    'chainType'?: string;
    /**
    * Name of the chain
    */
    'name': string;
    /**
    * The native coin of the chain
    */
    'coin': string;
    /**
    * Unique id of the chain
    */
    'id': number;
    /**
    * Whether the chain is mainnet or not
    */
    'mainnet': boolean;
    /**
    * Logo of the chain
    */
    'logoURI'?: string;
    /**
    * Url to the list of available tokens
    */
    'tokenlistUrl'?: string;
    /**
    * List of available faucets
    */
    'faucetUrls'?: Array<string>;
    /**
    * The multicall contract address
    */
    'multicallAddress'?: string;
    'metamask'?: ChainMetamask;
    'nativeToken'?: Token;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{ name: string, baseName: string, type: string, format: string }> = [
        {
            "name": "key",
            "baseName": "key",
            "type": "string",
            "format": ""
        },
        {
            "name": "chainType",
            "baseName": "chainType",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "coin",
            "baseName": "coin",
            "type": "string",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": "number"
        },
        {
            "name": "mainnet",
            "baseName": "mainnet",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "logoURI",
            "baseName": "logoURI",
            "type": "string",
            "format": ""
        },
        {
            "name": "tokenlistUrl",
            "baseName": "tokenlistUrl",
            "type": "string",
            "format": ""
        },
        {
            "name": "faucetUrls",
            "baseName": "faucetUrls",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "multicallAddress",
            "baseName": "multicallAddress",
            "type": "string",
            "format": ""
        },
        {
            "name": "metamask",
            "baseName": "metamask",
            "type": "ChainMetamask",
            "format": ""
        },
        {
            "name": "nativeToken",
            "baseName": "nativeToken",
            "type": "Token",
            "format": ""
        }];

    static getAttributeTypeMap() {
        return Chain.attributeTypeMap;
    }

    public constructor() {
    }
}