// lib/li.fi-ts/models/ChainMetamaskNativeCurrency.ts
/**
 * LI.FI API
 * LI.FI provides the best cross-chain swap across all liquidity pools and bridges.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ChainMetamaskNativeCurrency {
    'name'?: string;
    'symbol'?: string;
    'decimals'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "symbol",
            "baseName": "symbol",
            "type": "string",
            "format": ""
        },
        {
            "name": "decimals",
            "baseName": "decimals",
            "type": "number",
            "format": "number"
        }    ];

    static getAttributeTypeMap() {
        return ChainMetamaskNativeCurrency.attributeTypeMap;
    }

    public constructor() {
    }
}