// lib/li.fi-ts/models/ChainsResponse.ts
/**
 * LI.FI API
 * LI.FI provides the best cross-chain swap across all liquidity pools and bridges.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Chain } from './Chain';
import { HttpFile } from '../http/http';

/**
* 
*/
export class ChainsResponse {
    'chains'?: Array<Chain>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{ name: string, baseName: string, type: string, format: string }> = [
        {
            "name": "chains",
            "baseName": "chains",
            "type": "Array<Chain>",
            "format": ""
        }];

    static getAttributeTypeMap() {
        return ChainsResponse.attributeTypeMap;
    }

    public constructor() {
    }
}