// lib/li.fi-ts/models/FeeBalances.ts
/**
 * LI.FI API
 * LI.FI provides the best cross-chain swap across all liquidity pools and bridges.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { TokenBalances } from './TokenBalances';
import { HttpFile } from '../http/http';

export class FeeBalances {
    /**
    * The id of the chain
    */
    'chainId'?: number;
    'tokenBalances'?: Array<TokenBalances>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{ name: string, baseName: string, type: string, format: string }> = [
        {
            "name": "chainId",
            "baseName": "chainId",
            "type": "number",
            "format": ""
        },
        {
            "name": "tokenBalances",
            "baseName": "tokenBalances",
            "type": "Array<TokenBalances>",
            "format": ""
        }];

    static getAttributeTypeMap() {
        return FeeBalances.attributeTypeMap;
    }

    public constructor() {
    }
}